use user logged in status for private route redirection logic
Originally PrivateRoute was being used to only hide SquadHubPage when
club context was not set. Now, it is being used to isolate pages that
need authentication from those that do not. Hence, using the
isUserLoggedIn check to invoke redirection from the former to the latter
(the Sign In page in this case)