update user auth pages to use a prefix of `auth` in their routes
Originally, the paths for the sign in and sign up pages were prefixed
with the root path. This meant that the home page and the login page had
the same route path. To avoid conflict, the auth pages are now being
prefixed with `auth`. Additionally, if access to any path (not requiring
authentication) not defined in the UserAuth component is attempted, the
user is redirected to the root route path.