properly define routes
Previously, the auth and non-auth pages were rendered conditionally on
the basis of the user data being successfully fetched, which in turn,
relied on the presence of a valid authToken. This logic has been
scrapped and instead, the presence of the authToken is directly being
leveraged to separate pages requiring authentication and those that do
not. To simplify the structure, at the root level, there are two route
for auth and non-auth pages with the individual pages nested inside each
component now.